decouple StaticMethodCallAnalyzer
